+++ android/gltfio-android/src/main/java/com/google/android/filament/gltfio/FilamentAsset.java
@@ -242,16 +242,36 @@ public class FilamentAsset {
     }
 
     /**
+     * Attaches the given skin to the given node, which must have an associated mesh with
+     * BONE_INDICES and BONE_WEIGHTS attributes.
+     *
+     * This is a no-op if the given skin index or target is invalid.
+     */
+    public void attachSkin(@IntRange(from = 0) int skinIndex, @Entity int target) {
+        nAttachSkin(getNativeObject(), skinIndex, target);
+    }
+
+    /**
+     * Attaches the given skin to the given node, which must have an associated mesh with
+     * BONE_INDICES and BONE_WEIGHTS attributes.
+     *
+     * This is a no-op if the given skin index or target is invalid.
+     */
+    public void detachSkin(@IntRange(from = 0) int skinIndex, @Entity int target) {
+        nDetachSkin(getNativeObject(), skinIndex, target);
+    }
+
+    /**
      * Gets the joint count at skin index in this asset.
      */
-    public int getJointCountAt(int skinIndex) {
+    public int getJointCountAt(@IntRange(from = 0) int skinIndex) {
         return nGetJointCountAt(getNativeObject(), skinIndex);
     }
 
     /**
      * Gets joints at skin index in this asset.
      */
-    public @NonNull @Entity int[] getJointsAt(int skinIndex) {
+    public @NonNull @Entity int[] getJointsAt(@IntRange(from = 0) int skinIndex) {
         int[] result = new int[getJointCountAt(skinIndex)];
         nGetJointsAt(getNativeObject(), skinIndex, result);
         return result;
@@ -344,6 +364,9 @@ public class FilamentAsset {
 
     private static native int nGetMorphTargetCount(long nativeAsset, int entity);
     private static native void nGetMorphTargetNames(long nativeAsset, int entity, String[] result);
+
+    private static native void nAttachSkin(long nativeAsset, int skinIndex, int entity);
+    private static native void nDetachSkin(long nativeAsset, int skinIndex, int entity);
 
     private static native void nGetBoundingBox(long nativeAsset, float[] box);
     private static native String nGetName(long nativeAsset, int entity);