+++ filament/backend/src/noop/NoopDriver.cpp
@@ -262,15 +262,6 @@ void NoopDriver::endRenderPass(int) {
 void NoopDriver::nextSubpass(int) {
 }
 
-void NoopDriver::setRenderPrimitiveBuffer(Handle<HwRenderPrimitive> rph,
-        Handle<HwVertexBuffer> vbh, Handle<HwIndexBuffer> ibh) {
-}
-
-void NoopDriver::setRenderPrimitiveRange(Handle<HwRenderPrimitive> rph,
-        PrimitiveType pt, uint32_t offset,
-        uint32_t minIndex, uint32_t maxIndex, uint32_t count) {
-}
-
 void NoopDriver::makeCurrent(Handle<HwSwapChain> drawSch, Handle<HwSwapChain> readSch) {
 }