+++ filament/src/RenderPrimitive.cpp
@@ -30,7 +30,6 @@ void FRenderPrimitive::init(backend::DriverApi& driver,
 
     assert_invariant(entry.materialInstance);
 
-    mHandle = driver.createRenderPrimitive();
     mMaterialInstance = upcast(entry.materialInstance);
     mBlendOrder = entry.blendOrder;
 
@@ -43,10 +42,8 @@ void FRenderPrimitive::init(backend::DriverApi& driver,
         auto const& ebh = vertexBuffer->getHwHandle();
         auto const& ibh = indexBuffer->getHwHandle();
 
-        driver.setRenderPrimitiveBuffer(mHandle, ebh, ibh);
-        driver.setRenderPrimitiveRange(mHandle, entry.type,
-                (uint32_t)entry.offset, (uint32_t)entry.minIndex, (uint32_t)entry.maxIndex,
-                (uint32_t)entry.count);
+        mHandle = driver.createRenderPrimitive(ebh, ibh, entry.type, (uint32_t)entry.offset,
+                (uint32_t)entry.minIndex, (uint32_t)entry.maxIndex, (uint32_t)entry.count);
 
         mPrimitiveType = entry.type;
         mEnabledAttributes = enabledAttributes;
@@ -55,32 +52,31 @@ void FRenderPrimitive::init(backend::DriverApi& driver,
 
 void FRenderPrimitive::terminate(FEngine& engine) {
     FEngine::DriverApi& driver = engine.getDriverApi();
-    driver.destroyRenderPrimitive(mHandle);
+    if (mHandle) {
+        driver.destroyRenderPrimitive(mHandle);
+    }
 }
 
 void FRenderPrimitive::set(FEngine& engine, RenderableManager::PrimitiveType type,
         FVertexBuffer* vertices, FIndexBuffer* indices, size_t offset,
         size_t minIndex, size_t maxIndex, size_t count) noexcept {
+
     AttributeBitset enabledAttributes = vertices->getDeclaredAttributes();
+
     auto const& ebh = vertices->getHwHandle();
     auto const& ibh = indices->getHwHandle();
 
     FEngine::DriverApi& driver = engine.getDriverApi();
 
-    driver.setRenderPrimitiveBuffer(mHandle, ebh, ibh);
-    driver.setRenderPrimitiveRange(mHandle, type,
+    if (mHandle) {
+        driver.destroyRenderPrimitive(mHandle);
+    }
+
+    mHandle = driver.createRenderPrimitive(ebh, ibh, type,
             (uint32_t)offset, (uint32_t)minIndex, (uint32_t)maxIndex, (uint32_t)count);
 
     mPrimitiveType = type;
     mEnabledAttributes = enabledAttributes;
-}
-
-void FRenderPrimitive::set(FEngine& engine, RenderableManager::PrimitiveType type, size_t offset,
-        size_t minIndex, size_t maxIndex, size_t count) noexcept {
-    FEngine::DriverApi& driver = engine.getDriverApi();
-    driver.setRenderPrimitiveRange(mHandle, type,
-            (uint32_t)offset, (uint32_t)minIndex, (uint32_t)maxIndex, (uint32_t)count);
-    mPrimitiveType = type;
 }
 
 } // namespace filament